fix: use uint64 for interface statistics counters
The /proc/net/dev as well as /sys/class/net/{iface}/statistics uses struct
rtnl_link_stats64, where all counters are __u64.

Previously, parsing these values into int64 caused errors for large counters:

    strconv.ParseInt: parsing "18446744073709551613": value out of range

This commit ensures that all counters are treated as uint64.

See: https://www.kernel.org/doc/html/latest/networking/statistics.html

+++ lxd/instance/drivers/driver_qemu_metrics.go
@@ -54,14 +54,14 @@ func (d *qemu) getQemuMetrics() (*metrics.MetricSet, error) {
 
 		for name, state := range networkState {
 			out.Network[name] = metrics.NetworkMetrics{
-				ReceiveBytes:    uint64(state.Counters.BytesReceived),
-				ReceiveDrop:     uint64(state.Counters.PacketsDroppedInbound),
-				ReceiveErrors:   uint64(state.Counters.ErrorsReceived),
-				ReceivePackets:  uint64(state.Counters.PacketsReceived),
-				TransmitBytes:   uint64(state.Counters.BytesSent),
-				TransmitDrop:    uint64(state.Counters.PacketsDroppedOutbound),
-				TransmitErrors:  uint64(state.Counters.ErrorsSent),
-				TransmitPackets: uint64(state.Counters.PacketsSent),
+				ReceiveBytes:    state.Counters.BytesReceived,
+				ReceiveDrop:     state.Counters.PacketsDroppedInbound,
+				ReceiveErrors:   state.Counters.ErrorsReceived,
+				ReceivePackets:  state.Counters.PacketsReceived,
+				TransmitBytes:   state.Counters.BytesSent,
+				TransmitDrop:    state.Counters.PacketsDroppedOutbound,
+				TransmitErrors:  state.Counters.ErrorsSent,
+				TransmitPackets: state.Counters.PacketsSent,
 			}
 		}
 	}
